Move neri_schneider to utils module
